Fix evaluation issues in puiseux
Several bugs are fixed in the case when the variables in the curve
are not `x` and `y`. Sage's evaluation syntax `f(x=a,y=b)` breaks
when the variables `x` and `y` are not the polynomial ring variables
`x` and `y`. (For example, setting x = R.gen() when R = QQ['z'].)

Additionally, a bug in almost_monicize() caused rational expressions
to not automatically reduce. Unfortunately, expr.reduce() does not
work in some cases so manual division by x**l is performed.